Mass edits to default coverage rates
IMPORTANT change.

Before this commit we still had weird issues with the cardiovascular
module, and this seems to be due to two things
- there were CVD prevention interventions which were still scaling up.
This was because they loaded up a scale-up scenario by default, which
went from 5% to 50% e.g. 5, 6, 7, 8
- the coverage of all risk factor interventions were set to 100 for lots
of levels. This seemed to stop changes in coverage from doing anything.
This also may have been causing UI issues that were stopping us from
editing the coverage levels