Unzipped and renamed contents of PJNZ files
Now have the unzipped contents of both PJNZ files and have renamed them
to be carels and mine.

Something that is slightly confusing, is that I expected the filenames
of the unzipped files to be the same, but they weren't. Carel's unzipped
files were named after the filename (e.g. AFG_h1_2024...) and mine were
named after my template (e.g. asthma_copd_test... or something similar).

This could happen for two reasons
1. When you resave a file in Spectrum, it renames these files (this is
plausible)
2. Carel edited the files using his own program, which renamed them
(this is also plausible, but less likely)