DPs differ, but probably not important
There are two ways that .DP differs
-,<Value>,Asia,1,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,

This line is the most important. One file has "Asia" the other doesn't
have "Asia". That difference seems strange. This is for TFR I believe. I
suspect this doesn't have any implications, but it's worth noting.

The second difference is the way that scientific notation is rendered.
One file hsa 1E-5 for example, the other has 1E-05. I suspect this has
no importance either, because if it did the whole model would behave
weirdly probably.