Meaningful NC differences
The NC diff is what we suspected was the most meaningful.

The coverages are different and structured differently.
However, there are also other differences that warrant serious
consideration.
The best way to know these differences is through meld or some other
diff viewer, as there are too many to list here.

The first thing we will do is copy the mine.NC and change the RF
coverages from carels.NC, and see if this produces the same output.
If it does, we can assume the other changes are harmless.